+++ packages/expo-camera/CHANGELOG.md
@@ -50,6 +50,7 @@ _This version does not introduce any user-facing changes._
 ### 💡 Others
 
 - Make the casing of `Barcode` consistent. ([#26900](https://github.com/expo/expo/pull/26900) by [@alanjhughes](https://github.com/alanjhughes))
+- On `Android`, requesting audio permissions was meant to be optional in the config plugin. ([#27365](https://github.com/expo/expo/pull/27365) by [@alanjhughes](https://github.com/alanjhughes))
 
 ## 14.0.4 — 2024-02-06